contest/serializers.py
from rest_framework import serializers

from django.contrib.auth.models import User
from .models import Participant

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'email')

class ParticipantSerializer(serializers.ModelSerializer):
    user = UserSerializer(read_only=True)

    class Meta:
        model = Participant
        fields = ('penalty', 'points', 'user')