requirements.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-nested-admin==3.1.3
idna==2.8
kombu==4.2.2.post1
python-monkey-business==1.0.0
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
channels==2.1.7
channels-redis==2.4.0
djangorestframework==3.9.4
markdown==3.1.1
django-filter==2.1.0